Refactor: move agtkTypes to utils
